+++ eyecite/helpers.py
@@ -104,7 +104,7 @@ def add_post_citation(citation: CaseCitation, words: Tokens) -> None:
     citation.full_span_end = citation.span()[1] + m.end()
     citation.metadata.pin_cite = clean_pin_cite(m["pin_cite"]) or None
     if m["pin_cite"]:
-        citation.pin_cite_end=citation.span()[1] + len(m["pin_cite"])
+        citation.pin_cite_end = citation.span()[1] + len(m["pin_cite"])
 
     citation.metadata.extra = (m["extra"] or "").strip() or None
     citation.metadata.parenthetical = process_parenthetical(m["parenthetical"])
@@ -184,6 +184,11 @@ def add_pre_citation(citation: FullCaseCitation, words: Tokens) -> None:
     )
     if not m:
         return
+
+    if m["pin_cite"]:
+        # if a pin cite occurs before the citation mark it down
+        start, end = m.span()
+        citation.pin_cite_start = citation.span()[0] - (end - start)
 
     citation.metadata.pin_cite = clean_pin_cite(m["pin_cite"]) or None
     citation.metadata.antecedent_guess = m["antecedent"]